refactor: reorganize code into subpackages
