+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/context/TPCCContext.java
@@ -1,5 +1,7 @@
-package hu.bme.mit.ftsrg.hypernate;
+package hu.bme.mit.ftsrg.hypernate.context;
 
+import hu.bme.mit.ftsrg.hypernate.middleware.LoggingStubMiddleware;
+import hu.bme.mit.ftsrg.hypernate.middleware.WriteBackCachedChaincodeStubMiddleware;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Deque;
