+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/entity/DataAccessException.java
@@ -1,4 +1,4 @@
-package hu.bme.mit.ftsrg.hypernate;
+package hu.bme.mit.ftsrg.hypernate.entity;
 
 public abstract class DataAccessException extends Exception {
