+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/entity/EntityExistsException.java
@@ -1,4 +1,4 @@
-package hu.bme.mit.ftsrg.hypernate;
+package hu.bme.mit.ftsrg.hypernate.entity;
 
 public class EntityExistsException extends DataAccessException {
