+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/entity/SerializableEntityBase.java
@@ -1,7 +1,8 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 
-package hu.bme.mit.ftsrg.hypernate;
+package hu.bme.mit.ftsrg.hypernate.entity;
 
+import hu.bme.mit.ftsrg.hypernate.util.JSON;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.nio.charset.StandardCharsets;
