+++ lib/src/main/java/hu/bme/mit/ftsrg/hypernate/util/JSON.java
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 
-package hu.bme.mit.ftsrg.hypernate;
+package hu.bme.mit.ftsrg.hypernate.util;
 
 import com.google.gson.Gson;
 import lombok.experimental.UtilityClass;
